# docs/openapi.yml
openapi: 3.1.0
info:
  title: Aegiscan API
  summary: Aegiscan API
  description: Aegiscan is the open source Tines / Splunk SOAR alternative. You can
    operate Aegiscan in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Aegiscan Founders
    email: founders@digi-trans.org
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
...
  /workflows:
    get:
      tags:
      - workflows
      summary: List Workflows
      description: List workflows.
      operationId: workflows-list_workflows
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          description: Filter workflows by tags
          title: Tag
        description: Filter workflows by tags
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowReadMinimal'
                title: Response Workflows-List Workflows
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - workflows
      summary: Create Workflow
      description: 'Create a new Workflow.


        Optionally, you can provide a YAML file to create a workflow.

        You can also provide a title and description to create a blank workflow.'
      operationId: workflows-create_workflow
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        content:
          multipart/form-data:
            schema:
              allOf:
              - $ref: '#/components/schemas/Body_workflows-create_workflow'
              title: Body
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowReadMinimal'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /workflows/{workflow_id}/tags:
    get:
      tags:
      - workflows
      summary: List Tags
      description: List all tags for a workflow.
      operationId: workflows-list_tags
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagRead'
                title: Response Workflows-List Tags
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - workflows
      summary: Add Tag
      description: Add a tag to a workflow.
      operationId: workflows-add_tag
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowTagCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workflows/{workflow_id}/tags/{tag_id}:
    delete:
      tags:
      - workflows
      summary: Remove Tag
      operationId: workflows-remove_tag
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: tag_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Tag Id
      - name: workflow_id
        in: path
        required: true
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /schedules:
    get:
      tags:
      - schedules
      summary: List Schedules
      operationId: schedules-list_schedules
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: workflow_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          - type: 'null'
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
                title: Response Schedules-List Schedules
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - schedules
      summary: Create Schedule
      description: Create a schedule for a workflow.
      operationId: schedules-create_schedule
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /schedules/{schedule_id}:
    get:
      tags:
      - schedules
      summary: Get Schedule
      description: Get a schedule from a workflow.
      operationId: schedules-get_schedule
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: schedule_id
        in: path
        required: true
        schema:
          type: string
          pattern: sch-[0-9a-f]{32}
          title: Schedule Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - schedules
      summary: Update Schedule
      description: Update a schedule from a workflow. You cannot update the Workflow
        Definition, but you can update other fields.
      operationId: schedules-update_schedule
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: schedule_id
        in: path
        required: true
        schema:
          type: string
          pattern: sch-[0-9a-f]{32}
          title: Schedule Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - schedules
      summary: Delete Schedule
      description: Delete a schedule from a workflow.
      operationId: schedules-delete_schedule
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: schedule_id
        in: path
        required: true
        schema:
          type: string
          pattern: sch-[0-9a-f]{32}
          title: Schedule Id
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /schedules/search:
    get:
      tags:
      - schedules
      summary: Search Schedules
      description: '**[WORK IN PROGRESS]** Search for schedules.'
      operationId: schedules-search_schedules
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: workspace_id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleSearch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
                title: Response Schedules-Search Schedules
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_workflows-create_workflow:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
        file:
          anyOf:
          - type: string
            format: binary
          - type: 'null'
          title: File
      type: object
      title: Body_workflows-create_workflow
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Schedule:
      properties:
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        id:
          type: string
          title: Id
        status:
          type: string
          title: Status
          default: online
        cron:
          anyOf:
          - type: string
          - type: 'null'
          title: Cron
        inputs:
          type: object
          title: Inputs
        every:
          type: string
          format: duration
          title: Every
          description: ISO 8601 duration string
        offset:
          anyOf:
          - type: string
            format: duration
          - type: 'null'
          title: Offset
          description: ISO 8601 duration string
        start_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start At
          description: ISO 8601 datetime string
        end_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End At
          description: ISO 8601 datetime string
        timeout:
          anyOf:
          - type: number
          - type: 'null'
          title: Timeout
          description: The maximum number of seconds to wait for the workflow to complete
        workflow_id:
          type: string
          format: uuid
          title: Workflow Id
      type: object
      required:
      - owner_id
      - every
      - workflow_id
      title: Schedule
    ScheduleCreate:
      properties:
        workflow_id:
          anyOf:
          - type: string
            pattern: wf_[0-9a-zA-Z]+
          - type: string
            pattern: wf-[0-9a-f]{32}
          title: Workflow Id
        inputs:
          anyOf:
          - type: object
          - type: 'null'
          title: Inputs
        cron:
          anyOf:
          - type: string
          - type: 'null'
          title: Cron
        every:
          type: string
          format: duration
          title: Every
          description: ISO 8601 duration string
        offset:
          anyOf:
          - type: string
            format: duration
          - type: 'null'
          title: Offset
          description: ISO 8601 duration string
        start_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start At
          description: ISO 8601 datetime string
        end_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End At
          description: ISO 8601 datetime string
        status:
          type: string
          enum:
          - online
          - offline
          title: Status
          default: online
        timeout:
          type: number
          title: Timeout
          description: The maximum number of seconds to wait for the workflow to complete
          default: 300
      type: object
      required:
      - workflow_id
      - every
      title: ScheduleCreate
    ScheduleSearch:
      properties:
        workflow_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Workflow Id
        limit:
          type: integer
          title: Limit
          default: 100
        order_by:
          type: string
          title: Order By
          default: created_at
        query:
          anyOf:
          - type: string
          - type: 'null'
          title: Query
        group_by:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Group By
        agg:
          anyOf:
          - type: string
          - type: 'null'
          title: Agg
      type: object
      title: ScheduleSearch
    ScheduleUpdate:
      properties:
        inputs:
          anyOf:
          - type: object
          - type: 'null'
          title: Inputs
        cron:
          anyOf:
          - type: string
          - type: 'null'
          title: Cron
        every:
          anyOf:
          - type: string
            format: duration
          - type: 'null'
          title: Every
          description: ISO 8601 duration string
        offset:
          anyOf:
          - type: string
            format: duration
          - type: 'null'
          title: Offset
          description: ISO 8601 duration string
        start_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start At
          description: ISO 8601 datetime string
        end_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End At
          description: ISO 8601 datetime string
        status:
          anyOf:
          - type: string
            enum:
            - online
            - offline
          - type: 'null'
          title: Status
      type: object
      title: ScheduleUpdate
...
    TagRead:
      properties:
        id:
          type: string
          format: uuid4
          title: Id
        name:
          type: string
          maxLength: 50
          minLength: 1
          title: Name
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
          description: Hex color code
      type: object
      required:
      - id
      - name
      title: TagRead
      description: Model for reading tag data with validation.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WorkflowReadMinimal:
      properties:
        id:
          type: string
          pattern: wf_[0-9a-zA-Z]+
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        status:
          type: string
          title: Status
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        version:
          anyOf:
          - type: integer
          - type: 'null'
          title: Version
        tags:
          anyOf:
          - items:
              $ref: '#/components/schemas/TagRead'
            type: array
          - type: 'null'
          title: Tags
        alias:
          anyOf:
          - type: string
          - type: 'null'
          title: Alias
        error_handler:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Handler
      type: object
      required:
      - id
      - title
      - description
      - status
      - icon_url
      - created_at
      - updated_at
      - version
      title: WorkflowReadMinimal
    WorkflowTagCreate:
      properties:
        tag_id:
          type: string
          format: uuid4
          title: Tag Id
      type: object
      required:
      - tag_id
      title: WorkflowTagCreate
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-aegiscan-service-key
...